+++ packages/ui-components/docs/changelog.md
@@ -6,6 +6,7 @@ title: Changelog
 
 -   Fixed an issue where `Datepicker` in controlled mode would sometimes end up in an infinite loop.
 -   Fixed an issue where `Datepicker` if range was given did not show end year in the select. 
+-   Fixed an issue where `Table` column resizing would not affect cell width.
 
 ## 1.0.1