+++ packages/ui-components/docs/changelog.md
@@ -2,7 +2,7 @@
 title: Changelog
 ---
 
-## NEXT
+## 1.1.0
 
 -   Fixed an issue where `Datepicker` in controlled mode would sometimes end up in an infinite loop.
 -   Fixed an issue where `Datepicker` if range was given did not show end year in the select.